Split user dao on resetTokenDao and userDao
