+++ src/main/java/com/ncedu/cheetahtest/dao/user/UserDao.java
@@ -11,16 +11,6 @@ public interface UserDao {
 
     User findUserByEmailAndPassword(String email, String password);
 
-    void saveToken(ResetToken myToken);
-
-    ResetToken findResetTokenByToken(String token);
-
-    void createToken(ResetToken myToken);
-
-    void makeTokenExpired(ResetToken resetToken);
-
-    ResetToken findResetTokenByUserId(int id);
-
     void changeUserPassword(ResetToken resetToken, String password);
 
     User findUserByToken(String token);
