+++ src/main/java/com/ncedu/cheetahtest/dao/user/UserDaoImpl.java
@@ -1,5 +1,6 @@
 package com.ncedu.cheetahtest.dao.user;
 
+import com.ncedu.cheetahtest.dao.resettoken.ResetTokenRowMapper;
 import com.ncedu.cheetahtest.entity.user.User;
 import com.ncedu.cheetahtest.entity.user.ResetToken;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -73,36 +74,6 @@ public class UserDaoImpl implements UserDao {
     }
 
     @Override
-    public void saveToken(ResetToken myToken) {
-
-        String sql = "UPDATE reset_token SET token = ?, expiry_date = ? WHERE user_id = ?";
-
-        jdbcTemplate.execute(sql, (PreparedStatementCallback<Boolean>) preparedStatement -> {
-            preparedStatement.setString(1, myToken.getToken());
-            preparedStatement.setTimestamp(2, new Timestamp(new Date().getTime() + ResetToken.getEXPIRATION()));
-            preparedStatement.setInt(3, myToken.getDeveloperId());
-
-            return preparedStatement.execute();
-        });
-    }
-
-    @Override
-    public ResetToken findResetTokenByToken(String token) {
-        String sql = "SELECT id, token, expiry_date, user_id FROM reset_token WHERE token = ?";
-
-        List<ResetToken> resetTokens = jdbcTemplate.query(
-                sql,
-                preparedStatement -> preparedStatement.setString(1, token),
-                new ResetTokenRowMapper());
-
-        if (resetTokens.size() == 1) {
-            return resetTokens.get(0);
-        }
-
-        return null;
-    }
-
-    @Override
     public void changeUserPassword(ResetToken resetToken, String password) {
         String sql = "UPDATE users SET password = ? WHERE id = ?";
 
@@ -133,42 +104,5 @@ public class UserDaoImpl implements UserDao {
         return null;
     }
 
-    @Override
-    public ResetToken findResetTokenByUserId(int id) {
-        String sql = "SELECT id, token, expiry_date, user_id FROM reset_token WHERE user_id = ?";
 
-        List<ResetToken> resetTokens = jdbcTemplate.query(sql,
-                                        preparedStatement -> preparedStatement.setInt(1, id),
-                                        new ResetTokenRowMapper());
-
-        if (resetTokens.size() == 1) {
-            return resetTokens.get(0);
-        }
-
-        return null;
-    }
-
-    @Override
-    public void createToken(ResetToken myToken) {
-        String sql = "INSERT INTO reset_token (token, expiry_date, user_id) VALUES (?,?,?)";
-
-        jdbcTemplate.execute(sql, (PreparedStatementCallback<Boolean>) preparedStatement -> {
-            preparedStatement.setString(1, myToken.getToken());
-            preparedStatement.setTimestamp(2, new Timestamp(new Date().getTime() + ResetToken.getEXPIRATION()));
-            preparedStatement.setInt(3, myToken.getDeveloperId());
-
-            return preparedStatement.execute();
-        });
-    }
-
-    @Override
-    public void makeTokenExpired(ResetToken resetToken) {
-        String sql = "UPDATE reset_token SET expiry_date = current_timestamp WHERE token = ?";
-
-        jdbcTemplate.execute(sql, (PreparedStatementCallback<Boolean>) preparedStatement -> {
-            preparedStatement.setString(1, resetToken.getToken());
-
-            return preparedStatement.execute();
-        });
-    }
 }
